+++ internal/codegen/common/public_view.go
@@ -28,7 +28,7 @@ func BuildPublicView(domain *model.Domain) (*PublicView, error) {
 		Actors:    filter(domain.Actors(), func(value *model.Actor) bool { return value.Pub }),
 		Resources: filter(domain.Resources(), func(value *model.Resource) bool { return value.Pub }),
 		Events:    filter(domain.Events(), func(value *model.Data) bool { return value.Pub }),
-		Services:  filter(domain.Services(), func(value *model.Service) bool { return value.Pub }),
+		Services:  filter(domain.Services(), func(value *model.Service) bool { return value.Public() }),
 	}
 	collectViewData(domain, view)
 	if err := validatePublicView(domain, view); err != nil {