+++ internal/parser/grammar/declaration.go
@@ -65,6 +65,7 @@ type Service struct {
 	Decorators []*Decorator
 	Pub        bool
 	Api        bool
+	Open       bool
 	Name       *Identifier       `parser:"@@"`
 	Sections   []*ServiceSection `parser:"\"{\" (Newline)* (@@ (Newline)*)* \"}\""`